Fixed a crash caused by Spy (and Scrying Pool) (Issue 597)
This was because Player is not serializable.  I considered making Player
serializable, then decided the current implementation would probably leak too
much information to clients if I sent the whole Player object - a client that
wanted to peak into an opponent's hand would be able to, for instance.  So I
just changed it to send the name - that's already stored on the server,
anyway, and not localized, so it's probably ok to send a String in this
instance.